+++ labeddit/src/components/FeedPage/index.js
@@ -1,23 +1,11 @@
 import React from "react";
 import { useHistory } from "react-router-dom";
-import useForm from "../../hooks/useForm";
 import usePosts from "../../hooks/usePosts";
 import { Helmet } from "react-helmet-async";
 import PostCard from "../PostCard";
-import styled from "styled-components";
-import TextField from "@material-ui/core/TextField";
-import Container from "@material-ui/core/Container";
 import { makeStyles } from "@material-ui/core/styles";
-import Button from "@material-ui/core/Button";
-import { createPost } from "../../functions/axios";
-import { FormControlLabel } from "@material-ui/core";
 import CircularProgress from "@material-ui/core/CircularProgress";
-
-const FormContainer = styled.form`
-  display: grid;
-  justify-items: center;
-  margin-top: 20px;
-`;
+import PostForm from "../PostForm";
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -32,19 +20,6 @@ const FeedPage = () => {
   const [posts, getPosts] = usePosts();
   const history = useHistory();
   const classes = useStyles();
-  const [form, handleFormChange, resetForm] = useForm({ text: "", title: "" });
-
-  const handleFormSubmit = async (event) => {
-    event.preventDefault();
-    const success = await createPost(form);
-    if (success) {
-      resetForm();
-      getPosts();
-      alert("Post criado com sucesso");
-    } else {
-      alert("Erro ao criar Post");
-    }
-  };
 
   return posts.length > 0 ? (
     <div>
@@ -52,31 +27,7 @@ const FeedPage = () => {
         <title>LabEddit</title>
       </Helmet>
 
-      <FormContainer onSubmit={handleFormSubmit}>
-        <TextField
-          value={form.text}
-          id="outlined-basic"
-          variant="outlined"
-          placeholder="Escrever post"
-          type="text"
-          name="text"
-          onChange={handleFormChange}
-        />
-
-        <TextField
-          value={form.title}
-          id="outlined-basic"
-          variant="outlined"
-          placeholder="Título"
-          type="text"
-          name="title"
-          onChange={handleFormChange}
-        />
-
-        <Button type="submit" variant="contained">
-          Postar
-        </Button>
-      </FormContainer>
+      <PostForm getPosts={getPosts} />
 
       {posts.map((post) => (
         <PostCard
